requirements.txt
# FCAR Framework - Python Dependencies
# Install: pip install -r requirements.txt

# Core ML
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
joblib>=1.3

# Optimization solver
pyomo>=6.7
highspy>=1.7

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Dashboard
streamlit>=1.30

# REST API (NFR04)
fastapi>=0.100
uvicorn>=0.25
pydantic>=2.0

# Statistical testing
scipy>=1.11

# PDF extraction (optional, for proposal text)
pdfplumber>=0.10

# Testing
pytest>=7.4
pytest-cov>=4.1
